Removed unused code and updated documentation
