+++ BUILD
@@ -10,17 +10,12 @@ go_library(
     importpath = "github.com/flakrimjusufi/grpc-with-rest",
     visibility = ["//visibility:private"],
     deps = [
-        "//database",
+        "//client:client_lib",
         "//helper",
-        "//models",
         "//proto",
         "@com_github_grpc_ecosystem_grpc_gateway_v2//runtime",
         "@com_github_joho_godotenv//:godotenv",
         "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//codes",
-        "@org_golang_google_grpc//status",
-        "@org_golang_google_protobuf//encoding/protojson",
-        "@org_golang_google_protobuf//types/known/timestamppb",
     ],
 )
