+++ client/BUILD.bazel
@@ -1,18 +1,19 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "client_lib",
-    srcs = ["credit_card_server.go"],
+    srcs = [
+        "credit_card.go",
+        "user.go",
+    ],
     importpath = "github.com/flakrimjusufi/grpc-with-rest/client",
     visibility = ["//visibility:private"],
     deps = [
+        "//database",
+        "//models",
         "//proto",
-        "@org_golang_google_grpc//:go_default_library",
+        "@org_golang_google_grpc//codes",
+        "@org_golang_google_grpc//status",
+        "@org_golang_google_protobuf//types/known/timestamppb",
     ],
-)
-
-go_binary(
-    name = "client",
-    embed = [":client_lib"],
-    visibility = ["//visibility:public"],
 )
