--- handlers/.env.example
@@ -1,10 +1,0 @@
-DB_USERNAME=
-DB_PASSWORD=
-DB_DATABASE=
-DB_HOSTNAME=
-DB_PORT=
-DB_TYPE=
-POSTGRES_PASSWORD=
-SERVER_HOST=
-GRPC_SERVER_PORT=
-GRPC_GATEWAY_SERVER_PORT=
