+++ helper/helper.go
@@ -1,11 +1,13 @@
 package helper
 
 import (
-	"github.com/spf13/viper"
+	"net/http"
 	"regexp"
+
+	"github.com/spf13/viper"
 )
 
-// AllowedOrigin - Used to add the cors in header
+// AllowedOrigin - checks the origin
 func AllowedOrigin(origin string) bool {
 	if viper.GetString("cors") == "*" {
 		return true
@@ -14,4 +16,20 @@ func AllowedOrigin(origin string) bool {
 		return true
 	}
 	return false
+}
+
+// Cors - adding the cors to header
+func Cors(h http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if AllowedOrigin(r.Header.Get("Origin")) {
+			w.Header().Set("Access-Control-Allow-Origin", r.Header.Get("Origin"))
+			w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PATCH, DELETE")
+			w.Header().Set("Access-Control-Allow-Headers",
+				"Accept, Content-Type, Content-Length, Accept-Encoding, Authorization, ResponseType")
+		}
+		if r.Method == "OPTIONS" {
+			return
+		}
+		h.ServeHTTP(w, r)
+	})
 }
