+++ models/data_struct.go
@@ -1,8 +1,9 @@
 package models
 
 import (
-	"github.com/jinzhu/gorm"
 	"time"
+
+	"github.com/jinzhu/gorm"
 )
 
 // User - handles the ORM of Users table and the response of gRPC services
