--- server/BUILD.bazel
@@ -1,20 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-
-go_library(
-    name = "server_lib",
-    srcs = ["http_server.go"],
-    importpath = "github.com/flakrimjusufi/grpc-with-rest/server",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//handlers",
-        "@com_github_joho_godotenv//:godotenv",
-        "@com_github_labstack_echo_v4//:echo",
-        "@com_github_labstack_echo_v4//middleware",
-    ],
-)
-
-go_binary(
-    name = "server",
-    embed = [":server_lib"],
-    visibility = ["//visibility:public"],
-)
